refactored permutation into a submodule
